<nth-lab-start> Initial state of another lab
Note that there are code changes between the
end of the previous lab and the beginning of
this lab. This is okay. We call this "fast forwarding".

We recommend explaining any meaningful code
changes to the learner at the beginning of
the lab instructions, describing why we have
fast-forwarded the codebase, and directing the
learner to review the changes before starting
the lab.